requirements.txt
alabaster==0.7.10
anaconda-client==1.6.5
anaconda-navigator==1.6.9
anaconda-project==0.8.0
asn1crypto==0.22.0
astor==0.6.2
astroid==1.5.3
astropy==2.0.2
autopep8==1.4.3
Babel==2.5.0
backports-abc==0.5
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
backports.weakref==1.0.post1
basemap==1.0.7
benepar==0.0.3
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==1.5.0
bokeh==0.12.10
boto==2.48.0
Bottleneck==1.2.1
bz2file==0.98
cachetools==2.1.0
cdecimal==2.3
certifi==2018.11.29
cffi==1.10.0
chainer==3.5.0
chardet==3.0.4
click==6.7
cloudpickle==0.4.0
clyent==1.2.2
colorama==0.3.9
conda==4.4.4
conda-build==3.0.27
conda-verify==2.0.0
configparser==3.5.0
contextlib2==0.5.5
cryptography==2.0.3
cycler==0.10.0
cymem==1.31.2
Cython==0.26.1
cytoolz==0.9.0.1
dask==0.15.3
datashape==0.5.4
de-core-news-sm==2.0.0
decorator==4.3.0
dill==0.2.7.1
distributed==1.19.1
docutils==0.14
en-core-web-sm==2.0.0
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
fastcluster==1.1.24
fastrlock==0.3
fasttext==0.8.22
filelock==2.0.12
Flask==0.12.2
Flask-Cors==3.0.3
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.1.1
gast==0.2.0
gevent==1.2.2
glob2==0.5
gmpy2==2.0.8
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
greenlet==0.4.12
grin==1.2.1
grpcio==1.10.0
gunicorn==19.9.0
gym==0.10.5
h5py==2.8.0
heapdict==1.0.0
html5lib==0.9999999
httplib2==0.11.3
hypothesis==2.0.0
icecream==1.3
idna==2.6
imageio==2.2.0
imagesize==0.7.1
ipaddress==1.0.18
ipykernel==4.6.1
ipython==5.4.1
ipython-genutils==0.2.0
ipywidgets==7.0.0
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jupyterlab==0.27.0
jupyterlab-launcher==0.4.0
lazy-object-proxy==1.3.1
linecache2==1.0.0
llvmlite==0.20.0
locket==0.2.0
lxml==4.1.0
Markdown==2.6.10
MarkupSafe==1.0
matplotlib==2.1.1
mccabe==0.6.1
mistune==0.7.4
mkl-fft==1.0.0
mkl-random==1.0.1
mock==2.0.0
Morfessor==2.0.4
mpmath==0.19
msgpack==0.5.6
msgpack-numpy==0.4.1
msgpack-python==0.5.4
multipledispatch==0.4.9
murmurhash==0.28.0
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
nltk==3.2.4
nose==1.3.7
notebook==5.0.0
numba==0.35.0+10.g143f70e90.dirty
numexpr==2.6.4
numpy==1.14.3
numpydoc==0.7.0
oauth2client==4.1.2
odo==0.5.1
olefile==0.44
opencv-python==3.4.0.12
openpyxl==2.4.8
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.2
partd==0.3.8
path.py==10.3.1
pathlib==1.0.1
pathlib2==2.3.0
patsy==0.4.1
pbr==3.1.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
pip-autoremove==0.9.1
pkginfo==1.4.1
plac==0.9.6
ply==3.10
polyglot==16.7.4
preshed==1.0.0
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.0
ptyprocess==0.5.2
py==1.4.34
pyang==1.7.5
pyangbind-brcd==0.6.14
pyasn1==0.4.4
pyasn1-modules==0.2.2
pybind==0.1.35
pybind11==2.2.4
pycairo==1.13.3
pycld2==0.31
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.6.0
pyglet==1.3.2
Pygments==2.2.0
PyICU==2.0.3
pylint==1.7.4
pyodbc==4.0.17
pyOpenSSL==17.2.0
pyparsing==2.2.0
pyrouge==0.1.3
PySocks==1.6.7
pysrt==1.1.1
pytest==3.2.1
python-dateutil==2.6.1
python-Levenshtein==0.12.0
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
regex==2017.4.5
requests==2.18.4
rope==0.10.5
rsa==3.4.2
ruamel-yaml==0.11.14
scandir==1.6
scikit-image==0.13.0
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
sling==1.0.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
spacy==2.0.12
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.4
SQLAlchemy==1.1.13
statsmodels==0.8.0
subprocess32==3.2.7
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
tensor2tensor==1.7.0
tensorboard==1.10.0
tensorboardX==0.8
tensorflow-gpu==1.10.0
tensorflow-tensorboard==0.4.0rc3
termcolor==1.1.0
terminado==0.6
testpath==0.3.1
thinc==6.10.3
toolz==0.9.0
torch==1.0.0.dev20180930
torch-encoding==0.5.2
torchtext==0.2.3
torchvision==0.2.0
tornado==4.5.2
tqdm==4.20.0
traceback2==1.4.0
traitlets==4.3.2
typing==3.6.2
ujson==1.35
unicodecsv==0.14.1
unittest2==1.1.0
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.0.2
word2vec==0.9.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
zict==0.1.3
